Improved the script and renamed to build_runshfile
